+++ Source/ACE.Server/Command/Handlers/AdminCommands.cs
@@ -860,9 +860,9 @@ namespace ACE.Server.Command.Handlers
         [CommandHandler("addspell", AccessLevel.Developer, CommandHandlerFlag.RequiresWorld, 1, "Adds the specified spell to your own spellbook.", "<spellid>")]
         public static void HandleAdd(Session session, params string[] parameters)
         {
-            if (Enum.TryParse(parameters[0], true, out Network.Enum.Spell spellId))
+            if (Enum.TryParse(parameters[0], true, out SpellId spellId))
             {
-                if (Enum.IsDefined(typeof(Network.Enum.Spell), spellId))
+                if (Enum.IsDefined(typeof(SpellId), spellId))
                     session.Player.LearnSpellWithNetworking((uint)spellId);
             }
         }