+++ changelog.md
@@ -1,5 +1,9 @@
 # ACEmulator Change Log
 
+### 2018-12-09
+[Ripley]
+* Move Spell enum from server to Entity and rename SpellId
+
 ### 2018-12-07
 [Ripley]
 * Added some properties so the SQL writers could annotate them for readability (PCAP imported properties, used for dev reference only)